storageccl: move benchmarks to bench_test.go
They need to live in the storageccl_test package for the next commit,
but the tests want to stay in storageccl so we don't have to export the
eval* methods. Do this in its own commit to make it easier to review.
