+++ pkg/ccl/storageccl/writebatch_test.go
@@ -11,14 +11,11 @@ package storageccl
 import (
 	"bytes"
 	"reflect"
-	"strconv"
 	"testing"
-	"time"
 
 	"golang.org/x/net/context"
 
 	"github.com/cockroachdb/cockroach/pkg/base"
-	"github.com/cockroachdb/cockroach/pkg/internal/client"
 	"github.com/cockroachdb/cockroach/pkg/keys"
 	"github.com/cockroachdb/cockroach/pkg/roachpb"
 	"github.com/cockroachdb/cockroach/pkg/storage"
@@ -26,11 +23,8 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/storage/engine/enginepb"
 	"github.com/cockroachdb/cockroach/pkg/testutils"
 	"github.com/cockroachdb/cockroach/pkg/testutils/serverutils"
-	"github.com/cockroachdb/cockroach/pkg/testutils/testcluster"
-	"github.com/cockroachdb/cockroach/pkg/util/encoding"
 	"github.com/cockroachdb/cockroach/pkg/util/hlc"
 	"github.com/cockroachdb/cockroach/pkg/util/leaktest"
-	"github.com/cockroachdb/cockroach/pkg/util/randutil"
 )
 
 func TestDBWriteBatch(t *testing.T) {
@@ -178,48 +172,5 @@ func TestWriteBatchMVCCStats(t *testing.T) {
 	}
 	if !reflect.DeepEqual(expectedStats, cArgs.Stats) {
 		t.Errorf("mvcc stats mismatch %+v != %+v", expectedStats, cArgs.Stats)
-	}
-}
-
-func BenchmarkWriteBatch(b *testing.B) {
-	rng, _ := randutil.NewPseudoRand()
-	const payloadSize = 100
-	v := roachpb.MakeValueFromBytes(randutil.RandBytes(rng, payloadSize))
-
-	ts := hlc.NewClock(hlc.UnixNano, time.Nanosecond).Now()
-	for _, numEntries := range []int{100, 1000, 10000, 100000} {
-		b.Run(strconv.Itoa(numEntries), func(b *testing.B) {
-			ctx := context.Background()
-			tc := testcluster.StartTestCluster(b, 3, base.TestClusterArgs{})
-			defer tc.Stopper().Stop(ctx)
-			kvDB := tc.Server(0).KVClient().(*client.DB)
-
-			id := keys.MaxReservedDescID + 1
-			var batch engine.RocksDBBatchBuilder
-
-			var totalLen int64
-			b.ResetTimer()
-			for i := 0; i < b.N; i++ {
-				b.StopTimer()
-				prefix := roachpb.Key(keys.MakeTablePrefix(uint32(id)))
-				id++
-				for j := 0; j < numEntries; j++ {
-					key := encoding.EncodeUvarintAscending(prefix, uint64(j))
-					v.ClearChecksum()
-					v.InitChecksum(key)
-					batch.Put(engine.MVCCKey{Key: key, Timestamp: ts}, v.RawBytes)
-				}
-				end := prefix.PrefixEnd()
-				repr := batch.Finish()
-				totalLen += int64(len(repr))
-				b.StartTimer()
-
-				if err := kvDB.WriteBatch(ctx, prefix, end, repr); err != nil {
-					b.Fatalf("%+v", err)
-				}
-			}
-			b.StopTimer()
-			b.SetBytes(totalLen / int64(b.N))
-		})
 	}
 }
